Setup Clean Architecture project structure
