--- template-csharp-dotnet.http
@@ -1,6 +1,0 @@
-@template_csharp_dotnet_HostAddress = http://localhost:5296
-
-GET {{template_csharp_dotnet_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
